Move get_absorbers()
