+++ salsa/__init__.py
@@ -6,12 +6,12 @@ Written by: Brendan Boyd boydbre1@msu.edu
 
 __version__ = "0.1.0"
 
-from salsa.absorber_extractor import AbsorberExtractor
+from salsa.absorber_extractor import AbsorberExtractor, get_absorbers
 
 from salsa.absorber_plotter import AbsorberPlotter
 
 from salsa.generate_light_rays import generate_lrays, construct_rays, random_sightlines
 
-from salsa.generate_catalog import generate_catalog, get_absorbers
+from salsa.generate_catalog import generate_catalog
 
 from salsa import utils
